computer_vision/pyproject.toml
[tool.poetry]
name = "plant_pathology"
version = "0.1.1"
description = ""
authors = ["Hravan <krzysztof.krenc@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
tensorflow-addons = "^0.16.1"
pandas = "^1.4.1"
scikit-learn = "^1.0.2"
opencv-python = "^4.5.5"
matplotlib = "^3.5.1"
dltk = {path = '../dltk', develop = true}

[tool.poetry.dev-dependencies]
ipykernel = "^6.9.2"
black = {version = "^22.1.0", allow-prereleases = true}
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
